+++ modules/graphql_breadcrumbs/graphql_breadcrumbs.info.yml
@@ -1,6 +1,7 @@
 name: GraphQL Breadcrumbs
 type: module
 description: Expose a route's breadcrumbs to GraphQL
+hidden: true
 package: GraphQL
 core: 8.x
 dependencies:
