+++ modules/graphql_core/src/Plugin/GraphQL/Fields/Routing/LinkUrl.php
@@ -1,30 +1,30 @@
 <?php
 
-namespace Drupal\graphql_breadcrumbs\Plugin\GraphQL\Fields;
+namespace Drupal\graphql_core\Plugin\GraphQL\Fields\Routing;
 
 use Drupal\graphql\Plugin\GraphQL\Fields\FieldPluginBase;
 use Youshido\GraphQL\Execution\ResolveInfo;
 use Drupal\Core\Link;
 
 /**
- * Retrieve a link text.
+ * Retrieve a link's route object.
  *
  * @GraphQLField(
- *   id = "link_text",
+ *   id = "link_url",
  *   secure = true,
- *   name = "text",
- *   type = "String",
+ *   name = "url",
+ *   type = "Url",
  *   types = {"Link"}
  * )
  */
-class Label extends FieldPluginBase {
+class LinkUrl extends FieldPluginBase {
 
   /**
    * {@inheritdoc}
    */
   public function resolveValues($value, array $args, ResolveInfo $info) {
     if ($value instanceof Link) {
-      yield $value->getText();
+      yield $value->getUrl();
     }
   }
