+++ modules/graphql_core/src/Plugin/GraphQL/Types/Routing/Link.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\graphql_breadcrumbs\Plugin\GraphQL\Types;
+namespace Drupal\graphql_core\Plugin\GraphQL\Types\Routing;
 
 use Drupal\graphql\Plugin\GraphQL\Types\TypePluginBase;
