+++ modules/graphql_core/tests/modules/graphql_breadcrumbs_test/graphql_breadcrumbs_test.routing.yml
